refactor: align internal svgo declaration files
