+++ src/svgo/v1/index.ts
@@ -1,5 +1,6 @@
 import type { error } from "../../errors";
-import type { SVGO, SVGOptimizer, SVGOptions } from "./types";
+import type { SVGOptimizer } from "../types";
+import type { SVGO, SVGOptions } from "./svgo-v1";
 
 import SvgoV1Wrapper from "./wrapper";
