+++ src/svgo/v1/svgo-v1.d.ts
@@ -1,6 +1,3 @@
-import type { error } from "../../errors";
-import type { SVGOptimizer } from "../types";
-
 type SVGO = new (options: SVGOptions) => {
   optimize(svg: string): {
     readonly data: string;
@@ -10,8 +7,6 @@ type SVGO = new (options: SVGOptions) => {
 type SVGOptions = unknown;
 
 export type {
-  error,
   SVGO,
-  SVGOptimizer,
   SVGOptions,
 };
