--- src/svgo/v2/types.d.ts
@@ -1,5 +1,0 @@
-import type { SVGOptimizer } from "../types";
-
-export type {
-  SVGOptimizer,
-};
